# docs/swagger.yaml
basePath: /
definitions:
...
  models.OrderItem:
    properties:
      created_at:
        type: string
      food_id:
        type: string
      id:
        type: string
      order_id:
        type: string
      order_item_id:
        type: string
      quantity:
        description: s- small , m- medium . L -large
        type: string
      unit_price:
        type: number
      updated_at:
        type: string
    required:
    - food_id
    - quantity
    - unit_price
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: CRUP api's for go backend
  title: go-backend
  version: "1.0"
paths:
...
  /orderitems:
    get:
      description: Returns an array of orderitems placed  from the orderitemcollection
        in restaurent database.
      produces:
      - application/json
      responses:
        "200":
          description: 'slice of orderitems '
          schema:
            items:
              $ref: '#/definitions/models.OrderItem'
            type: array
      security:
      - Security.require(true): []
      summary: Returns slice of orderitems
      tags:
      - orderitem
    post:
      consumes:
      - application/json
      description: Creates a new orderitem resource on the server
      parameters:
      - description: Orderitem object
        in: body
        name: orderitme
        required: true
        schema:
          $ref: '#/definitions/models.OrderItem'
      produces:
      - application/json
      responses:
        "201":
          description: New orderitem created
          schema:
            $ref: '#/definitions/models.OrderItem'
        "500":
          description: Internal Server Error while creating a new orderitem
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Creates a orderitem  resource
      tags:
      - orderitem
  /orderitems/{orderitem_id}:
    get:
      consumes:
      - application/json
      description: Retrieves a orderitem with specific orderitem id from the orderitem
        collection
      parameters:
      - description: orderitem_id
        in: path
        name: orderitem_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Details of a specific orderitem id
          schema:
            $ref: '#/definitions/models.OrderItem'
        "500":
          description: Internal Server Error in mongodb
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Retrieves a orderitem with specific orderitem id
      tags:
      - orderitem
    patch:
      consumes:
      - application/json
      description: Updates an existing orderitem resource in the orderitem collection
      parameters:
      - description: ID of the orderitem  resource to update
        in: path
        name: orderitem_id
        required: true
        type: string
      - description: Orderitem object
        in: body
        name: orderitme
        required: true
        schema:
          $ref: '#/definitions/models.OrderItem'
      produces:
      - application/json
      responses:
        "200":
          description: orderitem  got updated with new body
          schema:
            $ref: '#/definitions/models.OrderItem'
        "404":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error while updating food
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Updates an oderitem resource
      tags:
      - orderitem
  /orderitemsorder/{order_id}:
    get:
      consumes:
      - application/json
      description: Retrieves orderitems placed from the specific order
      parameters:
      - description: order_id
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: retrieves orderitems placed from specific orderid
          schema:
            $ref: '#/definitions/models.OrderItem'
        "500":
          description: Internal Server Error in mongodb
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Retrieves orderitems placed from the specific order
      tags:
      - orderitem
...
securityDefinitions:
  "":
    description: JWT token for authentication
    in: header
    name: token
    type: apiKey
swagger: "2.0"